% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analises_sentinel.R
\name{analises_sentinel}
\alias{analises_sentinel}
\title{Função gera informações a partir de imagens sentinel já corrigidas (Download feito pelo plugin SCP - Semi-Automatic Classification plugin). Gera os seguintes indices de vegetação: NDVI, SFDVI, NDII e EVI. Gera também arquivo xlsx da mediana destes indices.}
\usage{
analises_sentinel(kml, past, savall)
}
\arguments{
\item{kml}{link dos arquivos kml}

\item{past}{link dos arquivos sentinel corrigidos}

\item{savall}{link da pasta para salvar todos os resultados}
}
\value{
Returns a images indices and data excel
}
\description{
Está função serve para analisar imagens do satélite sentinel
}
\author{
Santos Henrique Brant Dias
}
