% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_day_BR.R
\name{clim_day_BR}
\alias{clim_day_BR}
\title{Função gera os dados climáticos diários das estações automaticas do Brasil}
\usage{
clim_day_BR(date)
}
\arguments{
\item{date}{uma data}
}
\value{
Returns a data.frame with the AWS data requested
}
\description{
Está função serve para baixar os dados climáticos de todas estações INMET Brasil
}
\author{
Santos Henrique Brant Dias
}
