% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_hour_BR.R
\name{clim_hour_BR}
\alias{clim_hour_BR}
\title{Função gera os dados climáticos horários das estações automaticas do Brasil em uma determinada data}
\usage{
clim_hour_BR(date)
}
\arguments{
\item{date}{uma data}
}
\value{
Returns a data.frame with the AWS data requested
}
\description{
Está função serve para baixar os dados climáticos de todas estações INMET Brasil em uma determinada data
}
\author{
Santos Henrique Brant Dias
}
