% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hour_selected.R
\name{data_hour_selected}
\alias{data_hour_selected}
\title{Função gera os dados climáticos horários de determinados intervalo de dias de determinadas estações automaticas do INMET Brasil}
\usage{
data_hour_selected(date_ini, date_fin, stations)
}
\arguments{
\item{date_ini}{uma data inicial no formato ano-mes-dia '2022-11-16'}

\item{date_fin}{uma data final no formato ano-mes-dia '2022-11-16'}

\item{stations}{Número(s) das estações de interesse de acordo com a função station_INMT()}
}
\value{
Returns a data.frame with the AWS data requested
}
\description{
Está função serve para baixar os dados climáticos horários de determinados intervalo de dias de determinadas estações automaticas do INMET Brasil
}
\author{
Santos Henrique Brant Dias
}
