% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_day_BR.R
\docType{data}
\name{date}
\alias{date}
\title{Função gera os dados diários das estações automaticas do Brasil}
\format{
An object of class \code{Date} of length 1.
}
\usage{
date
}
\arguments{
\item{date}{uma data}
}
\value{
Returns a data.frame with the AWS data requested
}
\description{
Está função serve para calcular a ETo FAO 56 no Brasil
}
\author{
Santos Henrique Brant Dias
}
\keyword{datasets}
