% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_values_coordinates.R
\name{extract_values_coordinates}
\alias{extract_values_coordinates}
\title{Função para extrair valores dos rasters utilizando coordenadas individuais ou em dataframe}
\usage{
extract_values_coordinates(rasterE, Latitude, Longitude)
}
\arguments{
\item{rasterE}{um stack raster}

\item{Latitude}{um conjunto de coordenadas ou coluna dataframe de Latitude}

\item{Longitude}{um conjunto de coordenadas ou coluna dataframe de Longitude}
}
\value{
Returns a data.frame with the AWS data requested
}
\description{
Esta função serve para extrair valores de raster
}
\author{
Santos Henrique Brant Dias
}
