% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_values_kml.R
\name{extract_values_kml}
\alias{extract_values_kml}
\title{Função para extrair valores dos rasters utilizando coordenadas individuais ou em dataframe}
\usage{
extract_values_kml(rasterE, kml)
}
\arguments{
\item{rasterE}{um stack raster}

\item{kml}{um arquivo kml com os pontos a serem extraidos}
}
\value{
Returns a data.frame with the AWS data requested
}
\description{
Esta função serve para extrair valores de raster
}
\author{
Santos Henrique Brant Dias
}
