% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{seasons}
\alias{seasons}
\title{Função lista todas as estações automaticas do INMET Brasil}
\usage{
seasons()
}
\value{
Returns a data.frame with the AWS data requested
}
\description{
Função lista todas as estações automaticas do INMET Brasil
}
\author{
Santos Henrique Brant Dias
}
