% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_INMET.R
\name{station_INMET}
\alias{station_INMET}
\title{Função lista todas as estações automaticas do INMET Brasil}
\usage{
station_INMET()
}
\value{
Returns a data.frame with the AWS data requested
}
\description{
Está função serve para listar todas as estações INMET Brasil
}
\author{
Santos Henrique Brant Dias
}
